Linear Optimizer/api/requirements.txt
# Linear Optimizer API Dependencies
# Для работы с базой данных Altawin и веб-сервиса

fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
fdb==2.0.2
pydantic==2.5.0

# Дополнительные зависимости для работы с БД
python-multipart==0.0.6 